pyproject.toml
[tool.poetry]
name = "kernel"
version = "0.1.0"
description = ""
authors = ["michaelvdays <michaelv@playwithmydata.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"
requests = "^2.26.0"
Flask = "^2.0.2"
lxml = "^4.6.3"
smartystreets-python-sdk = "^4.10.2"
xlrd = "^2.0.1"
opentelemetry-instrumentation-flask = "^0.25b0"
opentelemetry-instrumentation-requests = "^0.25b0"
opentelemetry-sdk = "^1.6.0"
idna = "^3.3"
dnspython = "^2.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"